refactor(error): retype ErrorBus on unified api.error.AppError
