+++ shared/src/commonMain/kotlin/com/calypsan/listenup/client/data/sync/push/PushSyncOrchestrator.kt
@@ -2,7 +2,7 @@ package com.calypsan.listenup.client.data.sync.push
 
 import com.calypsan.listenup.client.core.error.ErrorBus
 import com.calypsan.listenup.client.core.toLegacy
-import com.calypsan.listenup.client.core.error.SyncError
+import com.calypsan.listenup.api.error.SyncError
 import com.calypsan.listenup.client.core.Success
 import com.calypsan.listenup.client.data.sync.SyncMutex
 import com.calypsan.listenup.client.domain.repository.NetworkMonitor
